\name{ScreenClean-package}
\alias{ScreenClean-package}
\alias{ScreenClean}
\docType{package}
\title{
Screen and clean variable selection procedures, including UPS and GS.
}
\description{Routines for a collection of screen-and-clean type variable selection procedures. 
}
\details{
\tabular{ll}{
Package: \tab ScreenClean\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-22\cr
License: \tab GPL (>= 2)\cr
}
}

\note{In order to use ScreenClean, the data need to be normalized, to make the standard deviation of the noise to be 1, and the l_2 norm of each length n predictor vector to be 1. }

\author{Pengsheng Ji, Jiashun Jin, Qi Zhang


Maintainer: Qi Zhang\email{qiz19@pitt.edu} 

}
\references{
Ji, P. and  Jin, J. (2012). \emph{UPS delivers optimal phase diagram in high dimensional variable selection.}
 Ann. Statist.,  40(1), 73-103.


Jin, J., Zhang, C.-H. and Zhang, Q. (2012). \emph{Optimality of Graphlet Screening in High Dimensional Variable Selection.} arXiv:1204.6452

}

