\name{VectorizeBase}
\alias{VectorizeBase}
\title{
expresses the number i on the base as a vector
}
\description{
expresses the number i on the base as a vector, an inner function.
}
\usage{
VectorizeBase(i, base, length)
}
\arguments{
  \item{i}{
the non-negative number to be converted
}
  \item{base}{the base to be converted on
}
  \item{length}{ the length of the converted vector
}
}

\value{
\item{vector}{A vector with the given length, whose elements can be read as the number i with the given base. }
}

